Builds final ThankYou component which is rendered after form submission. Styling. Form is functional.
