--- src/pages/landingPages/participate/ThankYou/ThankYou.js
@@ -1,14 +1,0 @@
-import "../Participate.scss";
-import Navigation from "../../../../components/navigation/mainNavMenu/navigation";
-import Footer from "../../../../components/footer/footer";
-// this page is under construction
-
-export default function ThankYou() {
-	return (
-		<div className="participate-landing-page">
-			<Navigation />
-			<h2>Application Submitted!</h2>
-			<Footer />
-		</div>
-	);
-}
