--- src/pages/landingPages/participate/components/ApplicationForm/ApplicationForm.scss
@@ -1,46 +1,0 @@
-.p-application-form {
-	width: 60%;
-	display: flex;
-	flex-direction: column;
-	gap: 1.1rem;
-	z-index: 3;
-
-	@media only screen and (max-width: 800px) {
-		width: 100%;
-	}
-
-	div {
-		display: flex;
-		flex-direction: column;
-		text-align: left;
-	}
-
-	input[type="text"],
-	input[type="email"],
-	input[type="url"],
-	select,
-	textarea {
-		background-color: rgb(236, 236, 236);
-		color: black;
-		border: 2px solid transparent;
-		border-radius: 6px;
-		padding: 2px 10px;
-		height: 2.3rem;
-		margin: 0;
-		display: block;
-
-		&:focus {
-			border: 2px solid var(--neon-purple);
-		}
-	}
-
-	textarea {
-		resize: none;
-		height: 6rem;
-		width: 100%;
-	}
-
-	select {
-		width: 50%;
-	}
-}
