+++ src/pages/landingPages/participate/components/Apply/Apply.scss
@@ -1,26 +1,104 @@
-.p-apply {
+.p-application-form {
+	width: 60%;
+	display: flex;
+	flex-direction: column;
+	gap: 1.1rem;
+	z-index: 3;
+
+	@media only screen and (max-width: 800px) {
+		width: 100%;
+	}
+
+	div {
+		display: flex;
+		flex-direction: column;
+		text-align: left;
+	}
+
+	input[type="text"],
+	input[type="email"],
+	input[type="url"],
+	select,
+	textarea {
+		background-color: rgb(236, 236, 236);
+		color: black;
+		border: 2px solid transparent;
+		border-radius: 6px;
+		padding: 2px 10px;
+		height: 2.3rem;
+		margin: 0;
+		display: block;
+
+		&:focus {
+			border: 2px solid var(--neon-purple);
+		}
+	}
+
+	textarea {
+		resize: none;
+		height: 6rem;
+		width: 100%;
+	}
+
+	select {
+		width: 50%;
+	}
+
+	label {
+		@media only screen and (max-width: 800px) {
+			font-size: 0.8rem;
+		}
+	}
+}
+
+.p-apply-thank-you {
 	position: relative;
 	display: flex;
 	flex-direction: column;
 	gap: 5rem;
-	z-index: 10;
 	min-height: 62vh;
 
 	h2 {
 		margin-top: 3rem;
+		font-size: 2rem;
+		z-index: 3;
 
-		@media only screen and (max-width: 800px) {
-			margin-top: 0;
+		@media only screen and (max-width: 1000px) {
 			font-size: 1.4rem;
+			margin: 0;
 		}
 	}
 
 	h3 {
-		font-size: 1.5rem;
+		font-size: 1.2rem;
+		letter-spacing: 1px;
+		padding: 0 1rem;
+		z-index: 3;
 
-		@media only screen and (max-width: 800px) {
-			font-size: 1rem;
+		@media only screen and (max-width: 1000px) {
+			font-size: 0.8rem;
+		}
+
+		a {
+			color: var(--neon-purple);
+		}
+	}
+
+	.learn-more {
+		font-size: 1.6rem;
+		z-index: 3;
+
+		@media only screen and (max-width: 1000px) {
+			font-size: 1.4rem;
+			margin: 0;
+		}
+	}
+
+	p {
+		@media only screen and (max-width: 1000px) {
+			font-size: 1.2rem;
 			line-height: 1.3;
+			padding: 0 2rem;
 		}
 	}
 
@@ -31,7 +109,7 @@
 		justify-content: center;
 		z-index: 3;
 
-		@media only screen and (max-width: 800px) {
+		@media only screen and (max-width: 1000px) {
 			flex-direction: column;
 			margin: 0 auto;
 		}
@@ -43,7 +121,7 @@
 			object-position: 100%;
 			margin-top: 28px;
 
-			@media only screen and (max-width: 800px) {
+			@media only screen and (max-width: 1000px) {
 				display: none;
 			}
 		}
@@ -53,18 +131,45 @@
 		display: flex;
 		gap: 5rem;
 
-		@media only screen and (max-width: 800px) {
+		@media only screen and (max-width: 1000px) {
 			gap: 2rem;
 		}
 
 		button {
 			width: 180px;
 
-			@media only screen and (max-width: 800px) {
+			@media only screen and (max-width: 1000px) {
 				width: 140px;
 				height: 55px;
 				margin-top: -55px;
 			}
+		}
+	}
+
+	/* this styling is to swap the order in which 
+	the elements are displayed on smaller screens */
+	.img-swap {
+		display: flex;
+		flex-direction: column;
+		justify-content: center;
+		align-items: center;
+		gap: 2rem;
+		width: 50vw;
+
+		@media only screen and (max-width: 1000px) {
+			flex-direction: column-reverse;
+			width: 100%;
+		}
+	}
+
+	.thank-you-img {
+		width: 100%;
+		height: 30rem;
+		object-fit: cover;
+		z-index: 3;
+
+		@media only screen and (max-width: 1000px) {
+			height: 15rem;
 		}
 	}
 
@@ -75,8 +180,8 @@
 	}
 
 	.green-planet {
-		bottom: 5%;
-		right: 5%;
+		bottom: 0%;
+		right: 0%;
 		width: 20rem;
 		transform: rotateY(180deg);
 
@@ -84,7 +189,7 @@
 			width: 12rem;
 		}
 
-		@media only screen and (max-width: 800px) {
+		@media only screen and (max-width: 1000px) {
 			display: none;
 		}
 	}
@@ -145,8 +250,8 @@
 
 	/* bottom left stars */
 	.star4 {
-		bottom: 20%;
-		left: 15%;
+		bottom: 2%;
+		left: 2%;
 		width: 2.4rem;
 		animation: twinkle 2s infinite;
 
@@ -162,8 +267,8 @@
 	}
 
 	.star5 {
-		bottom: 14%;
-		left: 23%;
+		bottom: 10%;
+		left: 10%;
 		width: 1.9rem;
 		animation: twinkle 2s infinite;
 		animation-delay: 0.2s;
